Refactor utility functions to api module
