+++ ros2bag/ros2bag/verb/record.py
@@ -12,60 +12,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import argparse
+from argparse import FileType
 import datetime
 import logging
 import os
-from typing import Dict
-from typing import Optional
 
-import rclpy
-from rclpy.duration import Duration
 from rclpy.qos import InvalidQoSProfileException
-from rclpy.qos import QoSProfile
+
 from ros2bag.verb import VerbExtension
 from ros2cli.node import NODE_NAME_PREFIX
 import yaml
 
-# This map needs to be updated when new policies are introduced
-POLICY_MAP = {
-    'history': rclpy.qos.QoSHistoryPolicy.get_from_short_key,
-    'reliability': rclpy.qos.QoSReliabilityPolicy.get_from_short_key,
-    'durability': rclpy.qos.QoSDurabilityPolicy.get_from_short_key,
-    'liveliness': rclpy.qos.QoSLivelinessPolicy.get_from_short_key
-}
+from ros2bag.api import create_bag_directory
+from ros2bag.api import validate_qos_profile_overrides
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger('ros2bag')
-
-
-def is_dict_valid_duration(duration_dict: Dict[str, int]) -> bool:
-    return all(key in duration_dict for key in ['sec', 'nsec'])
-
-
-def dict_to_duration(time_dict: Optional[Dict[str, int]]) -> Duration:
-    if time_dict:
-        if is_dict_valid_duration(time_dict):
-            return Duration(seconds=time_dict.get('sec'), nanoseconds=time_dict.get('nsec'))
-        else:
-            raise ValueError(
-                'Time overrides must include both seconds (sec) and nanoseconds (nsec).')
-    else:
-        return Duration()
-
-
-def validate_qos_profile_overrides(qos_profile_dict: Dict) -> Dict[str, Dict]:
-    """Validate the QoS profile yaml file path and its structure."""
-    for name in qos_profile_dict.keys():
-        profile = qos_profile_dict[name]
-        # Convert dict to Duration. Required for construction
-        conversion_keys = ['deadline', 'lifespan', 'liveliness_lease_duration']
-        for k in conversion_keys:
-            profile[k] = dict_to_duration(profile.get(k))
-        for policy in POLICY_MAP.keys():
-            profile[policy] = POLICY_MAP[policy](profile.get(policy, 'system_default'))
-        qos_profile_dict[name] = QoSProfile(**profile)
-    return qos_profile_dict
 
 
 class RecordVerb(VerbExtension):
@@ -122,16 +84,10 @@ class RecordVerb(VerbExtension):
             help='record also hidden topics.'
         )
         parser.add_argument(
-            '--qos-profile-overrides-path', type=argparse.FileType('r'),
+            '--qos-profile-overrides-path', type=FileType('r'),
             help='Path to a yaml file defining overrides of the QoS profile for specific topics.'
         )
         self._subparser = parser
-
-    def create_bag_directory(self, uri):
-        try:
-            os.makedirs(uri)
-        except OSError:
-            return "[ERROR] [ros2bag]: Could not create bag folder '{}'.".format(uri)
 
     def main(self, *, args):  # noqa: D102
         if args.all and args.topics:
@@ -156,7 +112,7 @@ class RecordVerb(VerbExtension):
                 logger.error(str(e))
                 return str(e)
 
-        self.create_bag_directory(uri)
+        create_bag_directory(uri)
 
         if args.all:
             # NOTE(hidmic): in merged install workspaces on Windows, Python entrypoint lookups
